refactor: update path to for other PRs
