+++ packages/angular/common/src/index.ts
@@ -1,2 +1,2 @@
 export { IonPopover } from './overlays/popover';
-export { ProxyCmp } from './utils';
+export { ProxyCmp } from './utils/proxy';
